Move router for clarity
